chore(dangerfile): seperate GitHub and GitLab stuff into seperate Dangefiles
In this way, this should fix the undefined BS stuff, finally.

BREAKING CHANGE: By default, Danger JS will look for the Dangerfile at root directory of the repo.
Since we move sutff into the danger directory, we need to add dangerfile flag when using 'danger
ci'. While this is not an breaking change and can be fixed in the next commit, this is just an note
to future maintainers and devs.

Signed-off-by: Andrei Jiroh Eugenio Halili <andreijiroh@madebythepins.tk>